issue 166 create template for notification

+++ core/src/main/java/com/epam/pipeline/entity/notification/NotificationTimestamp.java
@@ -14,7 +14,7 @@ public class NotificationTimestamp {
     private NotificationType type;
     private LocalDateTime timestamp;
 
-    public static boolean isTimeOutEnds(NotificationTimestamp nt, int shift) {
+    public static boolean isTimeOutEnds(NotificationTimestamp nt, long shift) {
         return nt != null && nt.getTimestamp() != null
                 && DateUtils.nowUTC().isAfter(nt.getTimestamp().plus(shift, ChronoUnit.MINUTES));
     }
